// src/utils/constants/enum.ts
export enum ResponseData {
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
}

export enum PaymentStatus {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  RETRYING = 'RETRYING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}

export enum PaymentCurrency {
  USD = 'USD',
  EUR = 'EUR',
  INR = 'INR',
  GBP = 'GBP',
}